+++ inst/doc/evaluating-rwa-method-reference.R
@@ -1,0 +1,11 @@
+## ----include = FALSE----------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%",
+  error = FALSE,
+  warning = FALSE,
+  message = FALSE
+)
+